backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
bidict==0.23.1
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
confluent-kafka==2.3.0
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.120.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
iniconfig==2.3.0
kafka-python==2.2.15
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyMySQL==1.1.0
pytest==7.4.3
python-dotenv==1.0.0
python-engineio==4.12.3
python-jose==3.5.0
python-socketio==5.11.0
PyYAML==6.0.3
redis==7.0.1
rsa==4.9.1
six==1.17.0
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.49.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wsproto==1.2.0
